[front] fix: split ConsumptionSummary from overview
